update page mappings
