+++ src/data/pageMappings.js
@@ -9,15 +9,16 @@ import { IsDefinedFunctor } from "../util/functors";
 // to emulate the behaviour of dynamic importing wrap the import statement with its string literal in a callback
 export const pageImportInfoCollection = Object.freeze({
     home: createPageImportInfo(() => import("../views/home-view.js")),
-    api: createPageImportInfo(() => import("../views/api-view.js"), api.pages),
-    documentation: createPageImportInfo(() => import("../views/documentation-view.js"), documentation.pages),
+    api: createPageImportInfo(() => import("../views/api-view.js"), api),
+    documentation: createPageImportInfo(() => import("../views/documentation-view.js"), documentation),
     download: createPageImportInfo(() => import("../views/download-view.js")),
     contacts: createPageImportInfo(() => import("../views/contacts-view.js")),
     search: createPageImportInfo(() => import("../views/search-view.js")),
 });
 
-export function tryGetSubpage(page, subpage) {
-    const ftor = IsDefinedFunctor(pageImportInfoCollection[page])
+export function tryGetSubpage(version, page, subpage) {
+    const ftor = IsDefinedFunctor(pageImportInfoCollection[version])
+        .map(version => version.pages[page])
         .map(pageInfo => pageInfo.subpages && pageInfo.subpages[subpage]);
 
     return ftor.valueOf()
@@ -25,21 +26,21 @@ export function tryGetSubpage(page, subpage) {
         : IsDefinedFunctor(getSubpageSuggestion(subpage));
 }
 
-function createPageImportInfo(importTargetCall, subpages = undefined) {
+function createPageImportInfo(importTargetCall, versionedSubpages = undefined) {
     return Object.freeze({
         importTarget: importTargetCall,
-        subpages,
-        requiresSubpage: !!subpages
+        versionedSubpages,
+        requiresSubpage: !!versionedSubpages
     });
 }
 
-function getSubpageSuggestion(subpage) {
-    if (api.pages[subpage]) {
-        return createSubpageSuggestion(`api`, api.pages[subpage]);
+function getSubpageSuggestion(version, subpage) {
+    if (api[version] && api[version].pages[subpage]) {
+        return createSubpageSuggestion(`api`, api[version].pages[subpage]);
     }
 
-    if (documentation.pages[subpage]) {
-        return createSubpageSuggestion(`documentation`, documentation.pages[subpage]);
+    if (documentation[version].pages[subpage]) {
+        return createSubpageSuggestion(`documentation`, documentation[version].pages[subpage]);
     }
 }
